added Files to set up for printing images using imageRecyclers
