--- app/src/main/java/com/example/mmm/SampleError.kt
@@ -1,8 +1,0 @@
-package com.example.mmm
-
-data class Error(
-    val userId: Int,
-    val id: Int,
-    val title: String,
-    val body: String
-)
