--- app/src/main/java/com/example/mmm/SampleResponse.kt
@@ -1,8 +1,0 @@
-package com.example.mmm
-
-data class Post(
-    val userId: Int,
-    val id: Int,
-    val title: String,
-    val body: String
-)
