+++ app/src/main/res/layout/activity_main.xml
@@ -52,6 +52,12 @@
                 android:paddingBottom="8dp"
                 android:paddingStart="8dp"
                 android:paddingEnd="8dp"/>
+
+            <androidx.recyclerview.widget.RecyclerView
+                android:id="@+id/recyclerView"
+                android:layout_width="match_parent"
+                android:layout_height="match_parent" />
+
             <ImageView
                 android:id="@+id/imageViewPosterHorror"
                 android:layout_width="wrap_content"
